refactor: move out tests
